+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `parent.onStream()` and `parent.createStream()` methods
+- `Service.createStream()` and `Service.onStream()` methods
+
 ## [0.2.2] - 2022-03-12
 
 ### Added